@@ -10,8 +10,7 @@
     "prebuild": "npm run clean",
     "build": "export NODE_ENV=production && node build.js",
     "clean": "rm -rf dist/",
-    "start": "node server.js",
-    "typings": "rm -rf typings/ && typings install"
+    "start": "node server.js"
   },
   "keywords": [
     "react",
@@ -19,52 +18,67 @@
     "typescript",
     "hot",
     "starter",
-    "seed"
+    "seed",
+    "webpack 2"
   ],
   "dependencies": {
-    "autoprefixer": "^6.3.3",
-    "awesome-typescript-loader": "^1.1.1",
+    "@types/classnames": "0.0.32",
+    "@types/lodash": "^4.14.52",
+    "@types/moment": "^2.13.0",
+    "@types/node": "^7.0.5",
+    "@types/react": "^15.0.12",
+    "@types/react-css-modules": "^3.7.6",
+    "@types/react-hot-loader": "^3.0.1",
+    "@types/react-redux": "^4.4.36",
+    "@types/react-router": "^3.0.4",
+    "@types/webpack-env": "^1.13.0",
+    "autoprefixer": "^6.7.5",
+    "awesome-typescript-loader": "^3.0.4",
     "babel-preset-react-hmre": "^1.1.1",
-    "classnames": "^2.2.3",
-    "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.25.0",
-    "cssnano": "^3.5.2",
-    "eslint": "^3.2.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "history": "^3.0.0",
-    "html-webpack-plugin": "^2.9.0",
-    "imports-loader": "^0.6.5",
-    "json-loader": "^0.5.4",
-    "postcss-cssnext": "^2.4.0",
-    "postcss-import": "^8.0.2",
-    "postcss-loader": "^0.13.0",
-    "postcss-modules-local-by-default": "^1.0.2",
+    "babel-core": "^6.23.1",
+    "babel-loader": "^6.3.2",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.22.0",
+    "classnames": "^2.2.5",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.26.1",
+    "cssnano": "^3.10.0",
+    "eslint": "^3.16.1",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "file-loader": "^0.10.0",
+    "history": "^4.5.1",
+    "html-webpack-plugin": "^2.28.0",
+    "imports-loader": "^0.7.0",
+    "postcss-cssnext": "^2.9.0",
+    "postcss-import": "^9.1.0",
+    "postcss-loader": "^1.3.1",
+    "postcss-modules-local-by-default": "^1.1.1",
     "postcss-nested": "^1.0.0",
     "raw-loader": "^0.5.1",
-    "react": "^15.3.0",
-    "react-css-modules": "^3.7.10",
-    "react-dom": "^15.3.0",
+    "react": "^15.4.2",
+    "react-css-modules": "^4.1.0",
+    "react-dom": "^15.4.2",
     "react-hot-loader": "^3.0.0-beta.4",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.5.2",
+    "react-redux": "^5.0.2",
+    "react-router": "^3.0.2",
     "redux": "^3.6.0",
-    "redux-logger": "^2.6.1",
-    "redux-thunk": "^2.1.0",
-    "source-map-support": "^0.4.0",
+    "redux-logger": "^2.8.1",
+    "redux-thunk": "^2.2.0",
+    "source-map-support": "^0.4.11",
     "style-loader": "^0.13.0",
-    "stylelint": "^7.4.0",
-    "stylelint-webpack-plugin": "^0.4.0",
-    "ts-helpers": "^1.1.1",
-    "tslint": "^3.7.4",
-    "tslint-loader": "^2.1.0",
-    "typescript": "^2.0.3",
-    "typings": "^1.0.4",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.12.0",
-    "webpack-split-by-path": "0.1.0-beta.1",
-    "whatwg-fetch": "^1.0.0"
+    "stylelint": "^7.9.0",
+    "stylelint-webpack-plugin": "^0.6.0",
+    "ts-helpers": "^1.1.2",
+    "tslint": "^4.4.2",
+    "tslint-loader": "^3.4.2",
+    "typescript": "^2.1.6",
+    "typings": "^2.1.0",
+    "typings-for-css-modules-loader": "^1.5.0",
+    "url-loader": "^0.5.7",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.4.1",
+    "webpack-hot-middleware": "^2.17.0",
+    "webpack-split-by-path": "2.0.0",
+    "whatwg-fetch": "^2.0.2"
   }
 }
