@@ -4,114 +4,10 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const StyleLintPlugin = require('stylelint-webpack-plugin');
 const SplitByPathPlugin = require('webpack-split-by-path');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const ForkCheckerPlugin = require('awesome-typescript-loader').ForkCheckerPlugin;
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 const IS_PRODUCTION = process.env.NODE_ENV === 'production';
 
-//
-// LOADERS
-//
-
-const loaders = {};
-
-loaders.js = {
-    test: /\.(jsx?)$/,
-    loaders: ['babel-loader?retainLines=true'],
-    exclude: /node_modules/
-};
-
-loaders.tsx = {
-    test: /\.(tsx?)$/,
-    loaders: ['awesome-typescript-loader?useBabel=true&useCache=true'],
-    exclude: /node_modules/
-};
-
-loaders.html = {
-    test: /\.html$/,
-    loader: 'raw',
-    exclude: /node_modules/
-};
-
-const localPostLoader = 'css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!postcss-loader';
-loaders.css = {
-    test: /\.css$/,
-    loader: IS_PRODUCTION ?
-        ExtractTextPlugin.extract('style-loader', localPostLoader) : ('style' + '!' + localPostLoader),
-    exclude: /(node_modules|\.global\.css)/
-};
-
-const globalPostLoader = 'css-loader!postcss-loader';
-loaders.globalcss = {
-    test: /\.global.css$/,
-    loader: IS_PRODUCTION ?
-        ExtractTextPlugin.extract('style-loader', globalPostLoader) : ('style' + '!' + globalPostLoader),
-    exclude: /node_modules/
-};
-
-loaders.json = {
-    test: /\.json$/,
-    loader: 'json'
-};
-
-loaders.ttfeot = {
-    test: /\.(ttf|eot)$/i,
-    loader: 'file?hash=sha512&digest=hex&name=fonts/[name].[ext]?[hash]'
-};
-
-loaders.woff = {
-    test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-    loader: 'url-loader?limit=10000&minetype=application/font-woff&name=fonts/[name].[ext]?[hash]'
-};
-
-loaders.image = {
-    test: /\.(jpe?g|png|gif|svg)$/i,
-    loader: 'file?hash=sha512&digest=hex&name=images/[name]-[hash].[ext]'
-};
-
-//
-// PLUGINS
-//
-
-const sourceMap = process.env.TEST || process.env.NODE_ENV !== 'production'
-  ? [new webpack.SourceMapDevToolPlugin({ filename: null, test: /\.tsx?$/ })]
-  : [];
-
-const basePlugins = [
-    new webpack.DefinePlugin({
-        'process.env': {
-            __DEV__: process.env.NODE_ENV !== 'production',
-            NODE_ENV: JSON.stringify(process.env.NODE_ENV),
-            BACKEND_URL: JSON.stringify(process.env.BACKEND_URL)
-        }
-    }),
-    new HtmlWebpackPlugin({
-        template: './src/index.html',
-        inject: 'body'
-    }),
-    new webpack.NoErrorsPlugin(),
-    new CopyWebpackPlugin([
-        { from: 'src/assets', to: 'assets' }
-    ]),
-].concat(sourceMap);
-
-const devPlugins = [
-    new StyleLintPlugin({
-        configFile: './.stylelintrc',
-        files: ['src/styles/*.css'],
-        failOnError: false
-    }),
-    new ForkCheckerPlugin(),
-    new webpack.HotModuleReplacementPlugin()
-];
-
-const prodPlugins = [
-    new ExtractTextPlugin('style.css', { allChunks: true })
-];
-
-const plugins = basePlugins
-  .concat(process.env.NODE_ENV === 'production' ? prodPlugins : [])
-  .concat(process.env.NODE_ENV !== 'production' ? devPlugins : []);
 
 //
 // POSTCSS
@@ -138,9 +34,171 @@ const postcssPlugins = postcssBasePlugins
   .concat(process.env.NODE_ENV === 'production' ? postcssProdPlugins : [])
   .concat(process.env.NODE_ENV === 'development' ? postcssDevPlugins : []);
 
-const postcssInit = () => {
-    return postcssPlugins;
+//
+// LOADERS
+//
+
+const loaders = {};
+
+loaders.js = {
+    test: /\.(jsx?)$/,
+    use: [{
+        loader: 'babel-loader',
+        options: {
+            retainLines: true
+        }
+    }],
+    exclude: /node_modules/
 };
+
+loaders.tsx = {
+    test: /\.(tsx?)$/,
+     use: [{
+        loader: 'awesome-typescript-loader',
+        options: {
+            useBabel: true,
+            useCache: true
+        }
+    }],
+    exclude: /node_modules/
+};
+
+loaders.html = {
+    test: /\.html$/,
+    use: ['raw-loader'],
+    exclude: /node_modules/
+};
+
+loaders.css = {
+    test: /\.css$/,
+    use: ExtractTextPlugin.extract({
+        fallback: 'style-loader',
+        use: [
+            {
+                loader: 'typings-for-css-modules-loader',
+                options: {
+                    modules: true,
+                    namedExport: true,
+                    camelCase: true
+                }
+            },
+            {
+                loader: 'postcss-loader',
+                options: {
+                    plugins: postcssPlugins,
+                    importLoaders: 1
+                }
+            }
+        ]
+    }),
+    exclude: /(node_modules|\.global\.css)/
+};
+
+loaders.globalcss = {
+    test: /\.global.css$/,
+    use: ExtractTextPlugin.extract({
+        fallback: 'style-loader',
+        use: [
+            {
+                loader: 'css-loader',
+                options: {
+                    localIdentName: '[name]'
+                }
+            },
+            {
+                loader: 'postcss-loader',
+                options: {
+                    plugins: postcssPlugins,
+                    importLoaders: 1
+                }
+            }
+        ]
+    }),
+    exclude: /node_modules/
+};
+
+exports.ttfeot = {
+    test: /\.(ttf|eot)$/i,
+    use: [{
+        loader: 'file-loader',
+        options: {
+            hash: 'sha512',
+            digest: 'hex',
+            name: 'fonts/[name].[ext]?[hash]'
+        }
+    }]
+};
+
+exports.woff = {
+    test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+    use: [{
+        loader: 'url-loader',
+        options: {
+            limit: 10000,
+            minetype: 'application/font-woff',
+            name: 'fonts/[name].[ext]?[hash]'
+        }
+    }]
+};
+
+exports.image = {
+    test: /\.(jpe?g|png|gif)$/i,
+    use: [{
+        loader: 'file-loader',
+        options: {
+            hash: 'sha512',
+            digest: 'hex',
+            name: 'images/[name]-[hash].[ext]'
+        }
+    }]
+};
+
+//
+// PLUGINS
+//
+
+const sourceMap = process.env.TEST || process.env.NODE_ENV !== 'production'
+  ? [new webpack.SourceMapDevToolPlugin({ filename: null, test: /\.tsx?$/ })]
+  : [];
+
+const basePlugins = [
+    new webpack.DefinePlugin({
+        'process.env': {
+            __DEV__: process.env.NODE_ENV !== 'production',
+            NODE_ENV: JSON.stringify(process.env.NODE_ENV),
+            BACKEND_URL: JSON.stringify(process.env.BACKEND_URL)
+        }
+    }),
+    new HtmlWebpackPlugin({
+        template: './src/index.html',
+        inject: 'body'
+    }),
+    new webpack.NoEmitOnErrorsPlugin(),
+    new CopyWebpackPlugin([
+        { from: 'src/assets', to: 'assets' }
+    ]),
+    new ExtractTextPlugin({
+        filename: 'style.css',
+        allChunks: true
+    })
+].concat(sourceMap);
+
+const devPlugins = [
+    new StyleLintPlugin({
+        configFile: './.stylelintrc',
+        files: ['src/styles/*.css'],
+        failOnError: false
+    }),
+    new webpack.HotModuleReplacementPlugin()
+];
+
+const prodPlugins = [
+    // new ExtractTextPlugin('style.css', { allChunks: true })
+];
+
+const plugins = basePlugins
+  .concat(process.env.NODE_ENV === 'production' ? prodPlugins : [])
+  .concat(process.env.NODE_ENV !== 'production' ? devPlugins : []);
 
 //
 // ENTRY
@@ -169,9 +227,11 @@ module.exports = {
       'inline-source-map',
 
     resolve: {
-        root: path.resolve('./'),
+        modules: [
+            path.resolve('./'),
+            'node_modules'
+        ],
         extensions: [
-            '',
             '.webpack.js',
             '.web.js',
             '.tsx',
@@ -188,15 +248,14 @@ module.exports = {
     },
 
     module: {
-        loaders: [
+        rules: [
             loaders.tsx,
-            loaders.html,
-            loaders.image,
+            // loaders.html,
+            // loaders.image,
             loaders.globalcss,
             loaders.css,
-            loaders.ttfeot,
-            loaders.woff,
-            loaders.json
+            // loaders.ttfeot,
+            // loaders.woff
         ]
     },
 
@@ -204,7 +263,5 @@ module.exports = {
         'react/lib/ReactContext': 'window',
         'react/lib/ExecutionEnvironment': true,
         'react/addons': true
-    },
-
-    postcss: postcssInit
+    }
 };
